Refactored the Action class.
Since the workflow for each feature implemented so far, and those planned for the future, is largely similar, there was a significant amount of duplicate code, and copying it each time was cumbersome. Therefore, an abstract base class `Action` was created, along with corresponding abstract base classes for `Preference`, `Params`, and `Generator`.

1.  The ①`Action`, ②`Generator`, and ③`Preference` classes each have generic type parameters:
    ① `<GeneratorType extends AbstractGenerator<ParamType>, PreferenceType extends AbstractPreference<ParamType>, ParamType extends AbstractParams>`
    ② `<ParamType extends AbstractParams>`
    ③ `<ParamType extends AbstractParams>`

2.  Concrete implementations of the abstract `Action` class must specify the generic types (e.g., the new `FilletAction` would specify `<FilletGenerator, FilletPreference, FilletParams>`).

3.  The constructor of a concrete `Action` class implementation needs to receive specific instances of `Preference` and `Generator`.