Changed the input parameters for the fillet tool. Now, it uses the more intuitive chainage interval to control smoothness, replacing the previous angular increment which was difficult to calculate.
Additionally, a recommended radius value is now provided. This radius is calculated to make the curves at two adjacent corners meet almost tangentially. Using a larger radius would cause the curves to intersect or fail to generate. (This recommendation is for guidance only; there is no enforced requirement for the input radius to be less than this value.)